Parsing HEC's UseACK by int, bool and string
Resolves #130 by parsing multiple different types of data in the
`useACK`field.

In our Cloud environment, this field comes as a stringified boolean,
such as a `"false"`. Which caused an error as the string wasn't able to
be parsed into an `int`.

These changes will cope with `useACK` as both `int` and `bool` by using
a custom json-marshaller ...

* `int` as a `string`: `"1"`
* `int` as an `int`: `1`
* `bool` as a `string`: `"true"`
* `bool` as a `bool`: `true`

There's test cases for all these variants.